backend/requirements.txt
# Flask and extensions
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
Flask-JWT-Extended==4.7.1
Flask-CORS==5.0.0
Flask-Limiter==3.8.0

# Database
SQLAlchemy==2.0.36

# Security
Werkzeug>=3.0.0
PyJWT>=2.8.0

# HTTP requests
requests==2.31.0

# Environment variables
python-dotenv==1.0.0

# Production server
gunicorn==21.2.0

# Development
pytest==7.4.3
pytest-cov==4.1.0
